pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "genlab"
version = "0.1.0"
description = "3DGenLab dry-run pipeline scaffold"
requires-python = ">=3.9"
dependencies = [
  "pyyaml",
  "numpy",
  "pillow",
  "trimesh",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
